feat: add certificate expiry alerts (#24)
We're now checking for certificate expiries in `uptime`, but we aren't
doing anything with the information aside from showing it in the
frontend. It would be good to notify through the same channels that we
are coming up to expiry so that we can renew them.

This change:
* Adds a `notification_type` table to denote whether something is an
`Uptime` notification or a `CertificateExpiry` notification
* Allows certificate expiries to notify every 3 days when fewer than 30
days are left
* Adds tests for these scenarios