feat: reduce log verbosity on miss (#28)
Since we are returning a `color_eyre::Report` from this method, the
debug output spans multiple lines which doesn't look great in logging.
There's actually only a single error case in this method (the `eyre!`
usage will never be hit and this can be tidied up in future) so let's
just return the `ResolveError` which will be shorter and have more
information.

This change:
* Swaps the logging to a warn as it is not an error usually
* Updates `resolve` to return the `ResolveError` instead of converting
it to a `Report`