refactor: allow `init` to do bootstrapping
Plenty of services will want to start up with access to a database, so
let's make this more of a feature in `init`.

Feature unification makes this a bit annoying in the workspace
(compiling stuff at the top level will enable more features than
compiling a specific project) but it's probably an ease of use
improvement.

This change:
* Adds a `database` feature to `foundation-init`
* Updates various projects to use it