feat: improve shutdown APIs
The `shutdown` APIs in `foundation/shutdown` are a bit difficult to work
with. We have certain services (like `uptime`) using both HTTP servers
and recurring jobs which can be tricky to shutdown properly.

Ideally, we'd just have a common way of doing things and make it easy
for multiple tasks to be spawned and awaited upon.

This change:
* Adds a `recurring-job` library in `foundation`
* Updates the `shutdown` API to use `CancellationToken` from
  `tokio_util` and tweaks the code to be easier to integrate with
* Updates all the usages