feat: trace configuration calls (#42)
Currently, `foundation-configuration` gets called before any of the
logging pipeline is set up as knowing whether we need OpenTelemetry
requires the configuration file to be parsed.

Instead, we can use `tracing`'s reloadable layers feature and allow the
configuration to be logged (but not sent to OpenTelemetry). Anything
that happens afterwards will support OpenTelemetry though.

This change:
* Tweaks the logging setup to allow the subscriber to be active earlier